Fixing Crashes
Fixed crashes due to missing and null values in the user data fields.

Also fixing a crash where it was asking for permissions when showing a notification in the background.

Lastly, Fixing a crash in the stats fragment which was caused by trying to get the stats from the db and display them even when the activity context was not added.